requirements.txt
streamlit>=1.30
pytest>=7.0
python-dateutil>=2.8
